Folder rename.
