+++ lib/screens/home/explorer/staking_delegation/staking_buttons.dart
@@ -1,6 +1,6 @@
 import 'package:provenance_wallet/common/pw_design.dart';
 import 'package:provenance_wallet/common/widgets/button.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_modal/staking_delegation_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_bloc.dart';
 import 'package:provenance_wallet/util/get.dart';
 
 class StakingButtons extends StatelessWidget {
