+++ lib/screens/home/explorer/staking_details/staking_details_screen.dart
@@ -3,9 +3,9 @@ import 'package:provenance_wallet/common/widgets/button.dart';
 import 'package:provenance_wallet/common/widgets/pw_app_bar.dart';
 import 'package:provenance_wallet/common/widgets/pw_list_divider.dart';
 import 'package:provenance_wallet/screens/home/explorer/explorer_bloc.dart';
+import 'package:provenance_wallet/screens/home/explorer/staking_delegation/staking_delegation_screen.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_details/details_item_copy.dart';
 import 'package:provenance_wallet/screens/home/explorer/staking_details/staking_details_bloc.dart';
-import 'package:provenance_wallet/screens/home/explorer/staking_modal/staking_delegation_screen.dart';
 import 'package:provenance_wallet/screens/home/transactions/details_item.dart';
 import 'package:provenance_wallet/services/account_service/account_service.dart';
 import 'package:provenance_wallet/services/account_service/transaction_handler.dart';
